% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coverage.R
\name{add_coverage}
\alias{add_coverage}
\title{Add coverage metrics to BLAST results}
\usage{
add_coverage(sample_id, blastres)
}
\arguments{
\item{sample_id}{character; sample ID}

\item{blastres}{data.frame; BLAST results}
}
\value{
data.frame; BLAST results with new columns
}
\description{
This function calculates coverage metrics based on BLAST results and adds
them to the output table.
}
