% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_qc.R
\name{check_qc}
\alias{check_qc}
\title{Check quality of assembled genomes}
\usage{
check_qc(df, taxid, new_na = "ignore", verbose = getOption(verbose))
}
\arguments{
\item{df}{data.frame; an assembly summary table.}

\item{taxid}{character; The NCBI Taxonomy ID for the organism. See Details
for more information.}

\item{new_na}{character; what should the function do if validation introduces
new NAs? Either \code{"ignore"} or \code{"stop"}.}

\item{verbose}{logical; Should verbose messages be printed to the console?}
}
\value{
A list with three elements:
\itemize{
  \item \code{"data"}: input data frame with new variable \code{qc_pass}. The
  new variable is \code{TRUE} if the assembly passed all QC checks, otherwise
  \code{FALSE}.
  \item \code{"failed"}: data frame of failed checks
  \item \code{"thresholds"}: a data frame of stochastic parameter thresholds
}
}
\description{
This function performs a number of deterministic and stochastic QC checks to
identify assemblies that may be of poor quality. Deterministic checks compare
the value of a variable with a fixed threshold, stochastic checks identify
statistically extreme values.
}
\details{
Some QC checks will run for each organism but some checks will only run for
selected organisms. For example, QC checks around capsular types will only
run for Acinetobacter baumannii or Klebsiella pneumoniae. This is because 
other organisms may not have capsules, or there may not be a typing scheme
for the capsule, or it is not yet implemented. Currently the following
\code{taxid} values have dedicated tests: \code{470} (Acinetobacter 
baumannii), \code{573} (Klebsiella pneumoniae).

Deterministic checks: \code{"mlst"} is not \code{NA}; exactly 7 genes
detected (470); \code{"k_serotype"} is not \code{NA} (470, 573); 
\code{"k_confidence"} is at least \code{"Good"} (470, 573); 
\code{"kraken2_taxid"} contains the taxid; \code{"coverage"} is at least 25; 
BUSCO complete score is at least 95%. Deterministic quality filters are 
applied once.

Stochastic checks: \code{"gc_content"} (use IQR extreme outliers, keep 
middle); \code{"contig_count"} (use IQR extreme outliers, keep lower); 
 \code{"longest_contig"} (use IQR extreme outliers, keep higher);
 \code{"N50"} (use IQR extreme outliers, keep higher); \code{"N95"} (use IQR
 extreme outliers, keep higher); \code{"n_count"} (use IQR extreme outliers, 
 keep lower, exclude 0); \code{"kraken2_taxid_freq"} (use IQR extreme
 outliers, keep higher). Stochastic quality checks are applied iteratively
 until no more rows are marked as outliers.
}
\examples{
\dontrun{
check_qc(df, taxid = 470, verbose = TRUE)
}
}
