% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citydict}
\alias{citydict}
\title{Translations of city names}
\format{
A data frame with 6 rows and 3 columns:
\describe{
  \item{international}{the international name of the city}
  \item{local}{the name of the city in a specific local context}
  \item{context}{the context in which the local name of the city may be used}
}
}
\usage{
citydict
}
\description{
This data frame currently contains Romanian cities which have separate names
in Hungarian. The data frame may be used during analysis to switch between
international and local names.
}
\keyword{datasets}
