% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{div}
\alias{div}
\title{Calculate diversity metrics}
\usage{
div(df, by, minyear, minyear_recent)
}
\arguments{
\item{df}{data.frame; a data frame which contains a row for each sample}

\item{by}{character; a variable of the data frame to be investigated}

\item{minyear}{numeric; the lowest year to include}

\item{minyear_recent}{numeric; the lowest year to include for calculations
on recent samples}
}
\value{
a data frame.
}
\description{
This function takes a data frame and a variable of the data frame and
calculates various diversity metrics e.g. Minor Allele Frequency (MAF) or
Shannon diversity.
}
