% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Downsample isolates}
\usage{
downsample(df, strategy, ...)
}
\arguments{
\item{df}{data.frame; a table with one row for each isolate}

\item{strategy}{character; the downsampling strategy, can be `geodate`,
`geodate2`, `geodate3`, `poppunk`, or `population`.}
}
\value{
The function returns a data frame which contains the same variables
as the input data frame but does not contain all isolates only the
downsampled isolates.
}
\description{
Downsample isolates using various downsampling strategies.
}
