% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_population.R
\name{downsample_population}
\alias{downsample_population}
\title{Downsample by population}
\usage{
downsample_population(aci, population, rate)
}
\arguments{
\item{aci}{data.frame; a table with one row for each sample.}

\item{population}{data.frame; a table with country names and populations.}

\item{rate}{numeric; sampling rate, samples per million inhabitants.}
}
\value{
A data frame which contains the same variables as the input data 
frame but does not contain all isolates only the downsampled isolates.
}
\description{
This function randomly selects isolates in proportion with the number of
inhabitants in each country.
}
