% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_orf_blast.R
\name{format_orf_blast}
\alias{format_orf_blast}
\title{Format ORF BLAST results}
\usage{
format_orf_blast(sample_id, orfs, blastres)
}
\arguments{
\item{sample_id}{character; sample ID}

\item{orfs}{data.frame; predicted ORFs parsed from a GFF file}

\item{blastres}{data.frame; BLAST results}
}
\value{
A formatted data frame.
}
\description{
This function uses ORF prediction and subsequent BLAST results to find BLAST
hit locations within genomes.
}
\details{
Formatting ORF BLAST results is a bit tricky because BLAST is called
on the Open Reading Frames (ORFs) and not the genomes. Because of this, query
locations will be locations within an ORF and not within a contig, and we
will never find hits on negative strands because those have already been
reverse complemented in the ORF prediction process. This function joins
results from ORF prediction and BLAST to trace results to locations within
the original genome (before ORF prediction).
}
