% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attribute.R
\name{get_attribute}
\alias{get_attribute}
\title{Get any attribute from a parsed GFF file}
\usage{
get_attribute(x, field, attrsep = ";")
}
\arguments{
\item{x}{character; attributes variable of the parsed gff file}

\item{field}{character; the attribute to be recovered}

\item{attrsep}{character; the separator which separates the attributes}
}
\value{
character; attribute values
}
\description{
GFF files include a field which contains a number of attributes concatenated
into a single variable. This function can be used to recover individual
attributes from this concatenated variable.
}
\note{
https://stat.ethz.ch/pipermail/bioconductor/2008-October/024669.html
}
\examples{
\dontrun {
# TODO include a working example
gff <- ape::read.gff("path_to_gff_file")
ID <- get_attribute(gff$attributes, "ID")
}
}
