% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_serofreq.R
\name{plot_serofreq}
\alias{plot_serofreq}
\title{Plot frequency of serotypes in data set}
\usage{
plot_serofreq(df, min_count = 1, max_bin_count = 3)
}
\arguments{
\item{df}{a data fram returned by the nextflow pipeline}

\item{min_count}{integer; the lowest ST-KL serogroup count to include}

\item{max_bin_count}{integer, the maximum number of resistance family bins
to include}
}
\description{
This function visualises the diversity of Multi Locus Sequence Types (MLSTs),
K Locus types and predicted resistance genes within the studied data set.
}
