% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot frequencies of character states over time}
\usage{
plot_ts(
  df,
  var,
  by = NULL,
  timevar = "collection_year",
  minyear = 2000,
  window = 2,
  type = "count",
  geom = "geom_line",
  show_count = TRUE,
  count_size = 8,
  colors = NULL
)
}
\arguments{
\item{df}{data.frame; a data frame}

\item{var}{character; the name of the character variable to plot}

\item{by}{character; the name of the character variable to group by}

\item{timevar}{character; the name of the variable which stores the year of
isolation. This variable must be numeric.}

\item{minyear}{numeric; samples isolated before this year will be removed
from the data set.}

\item{window}{numeric; number of preceding years to aggregate.}

\item{type}{character; defines transformation for the y axis. Can be either
\code{"count"} or \code{"ratio"}.}

\item{geom}{character; defines the plot type. Can be \code{"geom_line"},
\code{"geom_col"} or \code{"geom_area"}.}

\item{show_count}{logical; show sample counts for each year.}

\item{count_size}{integer; text size when showing counts for each year.}

\item{colors}{data.frame; a data frame for colors to use for color fill.
If \code{NULL}, the colors will be chosen automatically. If specified, the
data frame must have two columns, the first variable must have the same name
as the \code{var} parameter, the other must be \code{"color"}.}
}
\description{
This is a convenience function for plotting frequencies (counts or ratios)
of character states over time.
}
\examples{
\dontrun{
plot_ts(df = aci, var = "continent")
}
}
