% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_serotop_tbl.R
\name{prep_serotop_tbl}
\alias{prep_serotop_tbl}
\title{Cumulative relative prevalence of most prevalent types}
\usage{
prep_serotop_tbl(df, type, group_by)
}
\arguments{
\item{df}{data.frame; a table which contains one row for each sample and at
at least two columns: the `type` of the sample and a grouping variable.}

\item{type}{character; the name of a variable within `df` we are interested
in.}

\item{group_by}{character; the name of the grouping variable.}
}
\value{
a data.frame
}
\description{
This function prepares a table about the number of most prevalent types which
account for a certain ratio of all isolates.
}
