% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_geocode.R
\name{query_geocode}
\alias{query_geocode}
\title{Query a geocode from Google API}
\usage{
query_geocode(query, cache_file = "geocodes", verbose = getOption("verbose"))
}
\arguments{
\item{cache_file}{character; the name of the cache file without the file
extension.}

\item{verbose}{logical; should a verbose output be printed on the console?}

\item{location}{character; a vector of locations.}
}
\description{
This function is a wrapper around ggmap::geocode(). The Google Geocoding API
is not a free service, therefore any queries towards the API should make use
of caching. This function implements such caching approach.
}
\examples{
\donttest{
query_geocode("Budapest")
}
}
