% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_df}
\alias{read_df}
\title{Read table}
\usage{
read_df(df)
}
\arguments{
\item{df}{character; a file path}
}
\value{
a data frame
}
\description{
This is a convenience function for importing table formats that are commonly
used in this package, i.e. csv, tsv, rds using a single function.
}
