% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_cov.R
\name{seq_cov}
\alias{seq_cov}
\title{Calculate sequencing coverage}
\usage{
seq_cov(file)
}
\arguments{
\item{file}{character; file path to the fasta file}
}
\value{
A list with two elements. The first element is a tibble and contains
coverage data for each contig. The second element is the overall coverage for
the whole assembly. Overall coverage is calculated as the weighted average of
contig coverage, weighted by contig length.
}
\description{
This function takes an assembly fasta file and attempts to calculate
sequencing coverage from contig headers.
}
