% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_results.R
\name{tidy_results}
\alias{tidy_results}
\title{Combine prediction results with metadata}
\usage{
tidy_results(dir)
}
\arguments{
\item{dir}{character; path to the directory which contains the metadata and
the prediction results in standardized format.}
}
\value{
combined data set in .rds and .tsv format.
}
\description{
This function is the final step in an analysis job. It takes the metadata
file, validates its content, takes the prediction results file, merges the
two and exports the combined data set in two formats with proper names.
}
\examples{
\dontrun{
tidy_results(<path to the directory where analysis data is stored>)
}
}
