% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_metadata.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Validate metadata}
\usage{
validate_metadata(df, keep_extra = FALSE, reorder = TRUE)
}
\arguments{
\item{df}{data.frame; a data.frame of sample metadata}

\item{keep_extra}{logical; if the data.frame contains extra variables which
will not be checked by this function, should these columns be kept in the
returned object? Note that this may result in merge errors.}

\item{reorder}{logical; should the function return variables in a specific
order? If extra variables are kept, these variables will be put to the end of
the data frame.}
}
\value{
If the data frame meets all the requirements, it will be converted
to an object of class \code{meta_df}. Subsequent elements of the data
analytical pipeline will only accept metadata as a \code{meta_df} object.
}
\description{
This function checks whether the data frame which contains the sample
metadata contains certain columns and whether these columns have the right
classes.
}
\details{
Some variables are required and the function will stop with an
error if any of these variables are missing or not provided appropriately.
Other variables are optional and the function will print a warning message
when any of these variables are missing or stop with an error if any of them
are not provided appropriately. The function will look for the following
variables and classes:
\itemize{
\item \code{assembly} (required) character
\item \code{coverage} (optional) character
\item \code{organism} (required) character
\item \code{collection_date} (optional) character
\item \code{collection_year} (optional) integer
\item \code{collection_day} (optional) Date
\item \code{continent} (optional) character
\item \code{region23} (optional) character
\item \code{country} (optional)  character
\item \code{country_iso2c} (optional) character
\item \code{city} (optional) character
\item \code{ward} (optional) character
\item \code{covid} (optional) logical
\item \code{assembly_source} (required) character
\item \code{in_lab} (optional) logical
\item \code{lon} (optional) numeric
\item \code{lat} (optional) numeric
\item \code{jobname} (required) character
\item \code{relative_path} (required) character
}
}
\note{
The purpose of this function is to ensure that results from analysing
multiple batches of sequences are compatible and can be merged easily.
}
