% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_typing.R
\name{validate_typing}
\alias{validate_typing}
\title{Validate a typing summary table}
\usage{
validate_typing(df, taxid, new_na = "ignore", verbose = getOption("verbose"))
}
\arguments{
\item{df}{data.frame; an assembly summary table.}

\item{taxid}{character; The NCBI Taxonomy ID for the organism.}

\item{new_na}{character; what should the function do if validation introduces
new NAs? Either \code{"ignore"} or \code{"stop"}.}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
a data frame
}
\description{
There are functions and scripts in the package that take a typing summary
table as input. In order for these to work properly, the summary table must 
contain certain variables and these must be of the right type. This function
validates such a table, checks for variables names and classes, attempts to
resolve any issues and stops with an error if necessary.
}
