% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_var}
\alias{validate_var}
\title{Validate variable}
\usage{
validate_var(
  df,
  varname,
  varclass,
  coerce_fun = NULL,
  new_na = "ignore",
  verbose = getOption(verbose)
)
}
\arguments{
\item{df}{data.frame}

\item{varname}{character; the name of the variable to look for}

\item{varclass}{character; the class the variable should have}

\item{coerce_fun}{function; a user defined function to be used for coercion
if necessary. If \code{NULL} the function will not attempt coercion but stop
with an error instead if the class is not appropriate}

\item{new_na}{character; what should the function do if coercion introduces
new NAs? Either \code{"ignore"} or \code{"stop"}.}

\item{verbose}{logical; should verbose messages be printed to the console?}
}
\value{
A data frame where the variable is of the right class
}
\description{
This is a convenience function for validating variables of a data.frame. The
function checks if a certain variable is in the data frame and whether it has
the right class. If the variable does not have the right class, the function
can attempt to coerce it using a user defined coercion function.
}
\examples{
\dontrun{
# ensure a variable is numeric
df <- validate_var(
  df = df,
  varname = "count",
  varclass = "numeric",
  coerce_fun = function(x) as.numeric(x),
  verbose = TRUE
  )
}

}
