% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_tsv}
\alias{write_tsv}
\title{Export a tables}
\usage{
write_tsv(data, file)
}
\arguments{
\item{data}{a table, e.g. data frame or matrix}

\item{file}{path to the output file}
}
\description{
This is a convenience function for exporting tables generated by any of the
processes.
}
